Conflicts:
	core/src/main/java/org/apache/accumulo/core/client/admin/TableOperations.java
	core/src/main/java/org/apache/accumulo/core/client/admin/TableOperationsImpl.java
	core/src/main/java/org/apache/accumulo/core/client/impl/ConnectorImpl.java
	core/src/main/java/org/apache/accumulo/core/client/mock/MockInstanceOperations.java
	core/src/main/java/org/apache/accumulo/core/iterators/FirstEntryInRowIterator.java
	core/src/main/java/org/apache/accumulo/core/iterators/IteratorUtil.java
	core/src/main/java/org/apache/accumulo/core/iterators/user/IntersectingIterator.java
	core/src/main/java/org/apache/accumulo/core/security/ColumnVisibility.java
	core/src/test/java/org/apache/accumulo/core/iterators/user/CombinerTest.java
	examples/simple/src/main/java/org/apache/accumulo/examples/simple/shard/Query.java
	minicluster/src/main/java/org/apache/accumulo/minicluster/MiniAccumuloCluster.java
	server/src/main/java/org/apache/accumulo/server/master/balancer/TabletBalancer.java
	server/src/main/java/org/apache/accumulo/server/util/AddFilesWithMissingEntries.java
	server/src/main/java/org/apache/accumulo/server/util/DumpZookeeper.java
	server/src/main/java/org/apache/accumulo/server/util/FindOfflineTablets.java
	server/src/main/java/org/apache/accumulo/server/util/MetadataTable.java
	server/src/main/java/org/apache/accumulo/server/util/RestoreZookeeper.java
	server/src/main/java/org/apache/accumulo/server/util/SendLogToChainsaw.java
	server/src/main/java/org/apache/accumulo/server/util/TableDiskUsage.java
	server/src/main/java/org/apache/accumulo/server/util/TabletServerLocks.java
	src/core/src/main/java/org/apache/accumulo/core/client/mapreduce/InputFormatBase.java
	src/core/src/main/java/org/apache/accumulo/core/file/map/MapFileUtil.java
	src/server/src/main/java/org/apache/accumulo/server/util/DumpMapFile.java
	src/server/src/main/java/org/apache/accumulo/server/util/DumpTabletsOnServer.java
	test/src/main/java/org/apache/accumulo/test/continuous/ContinuousMoru.java
	test/src/main/java/org/apache/accumulo/test/randomwalk/concurrent/CheckBalance.java
	trace/src/main/java/org/apache/accumulo/trace/instrument/receivers/ZooSpanClient.java

+++ test/src/main/java/org/apache/accumulo/test/randomwalk/concurrent/CheckBalance.java
@@ -37,9 +37,6 @@ public class CheckBalance extends Test {
   static final String LAST_UNBALANCED_TIME = "lastUnbalancedTime";
   static final String UNBALANCED_COUNT = "unbalancedCount";
 
-  /* (non-Javadoc)
-   * @see org.apache.accumulo.test.randomwalk.Node#visit(org.apache.accumulo.test.randomwalk.State, java.util.Properties)
-   */
   @Override
   public void visit(State state, Properties props) throws Exception {
     log.debug("checking balance");
@@ -105,6 +102,6 @@ public class CheckBalance extends Test {
       double diff = s.doubleValue() - avg;
       sqrtotal += diff * diff;
     }
-    return Math.sqrt(sqrtotal / (double) num);
+    return Math.sqrt(sqrtotal / num);
   }
 }
